+++ backend/src/main/java/com/terradev/cleanworld/service/ZoneService.java
@@ -17,18 +17,42 @@ public class ZoneService {
         this.repository = repository;
     }
 
+    /**
+     * GET -> Obtener todas las zonas
+     *
+     * @return Lista de todas las zonas
+     */
     public List<ZoneEntity> findAll() {
         return repository.findAll();
     }
 
+    /**
+     * GET -> Obtener unazona por ID
+     *
+     * @param id Identificador de la zona
+     * @return Optional con la zona indicado por ID
+     */
     public Optional<ZoneEntity> findById(Long id) {
         return repository.findById(id);
     }
 
+    /**
+     * POST -> Crear una nueva zona
+     *
+     * @param z Entidad ZoneEntity con los datos a guardar
+     * @return Zona Creada
+     */
     public ZoneEntity save(ZoneEntity z) {
         return repository.save(z);
     }
 
+    /**
+     * PATCH -> Actualizar parcialmente una zona
+     *
+     * @param id Identificador de la zona
+     * @param updates Campos que se desean actualizar
+     * @return Zona actualizada
+     */
     public Optional<ZoneEntity> patchZone(Long id, Map<String, Object> updates) {
         return repository.findById(id).map(existing -> {
 
@@ -56,7 +80,16 @@ public class ZoneService {
         });
     }
 
+    /**
+     * DELETE -> Eliminar zona
+     *
+     * @param id Identificador de la zona a eliminar
+     * @throws RuntimeException Si la zona no existe
+     */
     public void deteleById(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Zona no encontrada con id " + id);
+        }
         repository.deleteById(id);
     }
 }
